Move convert into its own package and introduce kubebuilder client
Split the convert subcommand out of pkg into a dedicated pkg/convert
package (convert.go for the command and I/O helpers, resources.go for
the Stash-to-KubeStash conversion logic). Wire up an uncached
controller-runtime client registering both the Stash (v1alpha1,
v1beta1) and KubeStash (storage, core) schemes.
